Fix GPL SPDX license identifier in Cargo manifests
According to
https://doc.rust-lang.org/cargo/reference/manifest.html#the-license-and-license-file-fields,
this field must be set to a known license from the SPDX license list
3.11. However, in such list, available at
  https://github.com/spdx/license-list-data/blob/v3.11/licenses.md,
there is no "GPL-3.0+" identifier. The "GPL-3.0-or-later" identifier
should be used instead.

This incorrect SPDX identifier makes tools such as cargo deny complain,
and cargo deny is pretty useful for checking that the dependency chain
fullfills expectations related to licensing and other characteristics.

These changes address the situation by replacing the "GPL-3.0+"
identifier with "GPL-3.0-or-later", as likely intended.